Reorganise fonts module (new submodules)
